+++ Gunman/src/main/java/entities/HealthEntity.java
@@ -21,8 +21,8 @@ public class HealthEntity extends Entity {
 
 	@Override
 	public boolean collidesWith(Entity other) {
-		hero.setBounds((int) x, (int) y, sprite.getWidth(), sprite.getHeight());
-		health.setBounds((int) other.x, (int) other.y, other.sprite.getWidth(), other.sprite.getHeight());
+		hero.setBounds(x, y, sprite.getWidth(), sprite.getHeight());
+		health.setBounds(other.x, other.y, other.sprite.getWidth(), other.sprite.getHeight());
 		if (hero.intersects(health))
 			return true;
 		return false;
